backend/requirements.txt
fastapi==0.115.6
python-dotenv==1.0.1
uvicorn[standard]==0.34.0
anthropic==0.42.0
httpx==0.28.1
python-multipart==0.0.20
pydantic==2.10.4
jinja2==3.1.5
google-auth==2.37.0
google-api-python-client==2.159.0
